demos: high_res: Clear /dev/fb0 on exit
When the exit button is pressed, the demo stops running, but the fbdev
is not cleared. This has the effect that the display still shows the
demo's last screen, as if it were still running.

To fix this, clear the fbdev screen in the `exit_cb` function callback,
and show a blank screen.

Signed-off-by: Suhaas Joshi <s-joshi@ti.com>